--- server/src/accounts/views.py
@@ -1,62 +1,0 @@
-from flask import request, jsonify 
-from accounts import app 
-from accounts.models import db, User
-
-#create user account
-@app.route('/users', methods=['POST'])
-def create_user():
-    data = request.get_json()
-    new_user = User(username=data['username'], email=data['email'], password=data['password'])
-    db.session.add(new_user)
-    db.session.commit()
-    return jsonify({'message': 'User account created successfully'}), 201
-
-#retrieve all user accounts
-@app.route('/users', methods=['GET'])
-def get_all_users():
-    users = User.query.all()
-    result = []
-    for user in users:
-        user_data = {
-            'id': user.id,
-            'username': user.username,
-            'email': user.email
-        }
-        result.append(user_data)
-    return jsonify(result), 200
-
-# Retrieve a single user ID 
-@app.route('/users/<int:user_id>', methods=['GET'])
-def get_user(user_id):
-    user = User.query.get(user_id)
-    if not user:
-        return jsonify({'message': 'User not found'}), 404
-    user_data = {
-        'id': user.id,
-        'username': user.username,
-        'email': user.email
-    }
-    return jsonify(user_data), 200
-
-# Update a single user account
-@app.route('/users/<int:user_id>', methods=['PUT'])
-def update_user(user_id):
-    user = User.query.get(user_id)
-    if not user:
-        return jsonify({'message': 'User not found'}), 404
-    data = request.get_json()
-    user.username = data.get('username', user.username)
-    user.email = data.get('email', user.email)
-    user.password = data.get('password', user.password)
-    db.session.commit()
-    return jsonify({'message': 'User updated successfully'}), 200
-
-# Delete a user account by ID
-@app.route('/users/<int:user_id>', methods=['DELETE'])
-def delete_user(user_id):
-    user = User.query.get(user_id)
-    if not user:
-        return jsonify({'message': 'User not found'}), 404
-    db.session.delete(user)
-    db.session.commit()
-    return jsonify({'message': 'User deleted successfully'}), 200
